third_party/blink/web_tests/platform/generic/fast/dom/option-properties-expected.txt

This page tests setting 'selected' and 'text' on an option element. See https://bugs.webkit.org/show_bug.cgi?id=9095.

If the test passes, you'll see a series of 'PASS' messages below.

PASS: document.getElementById('sel').options.selectedIndex should be 1 and is.
PASS: document.getElementById('sel').options[1].selected should be true and is.
PASS: document.getElementById('sel').options[1].text should be PASS and is.
FAIL
PASS
